Refactored arbitrage
