+++ server/server/urls.py
@@ -22,6 +22,6 @@ urlpatterns = [
     path('insights/', include('ai_insights.urls')),
     path('core/', include('core.urls')),
     path('odds/', include('odds.urls')),
-    path('arbitrage/', include('arbitrage.urls'))
+    path("arbitrage/", include("odds.arbitrage.urls"))
 
 ]
